+++ projects/composablekernel/include/ck/tensor_operation/gpu/device/impl/device_grouped_conv_fwd_multiple_abd_xdl_cshuffle.hpp
@@ -106,11 +106,11 @@ __global__ void
     const index_t g_idx = __builtin_amdgcn_readfirstlane(blockIdx.y);
     const index_t n_idx = __builtin_amdgcn_readfirstlane(blockIdx.z);
     const long_index_t e_group_offset =
-        amd_wave_read_first_lane(int64_t(compute_ptr_offset_of_groups.GetEPtrOffset(g_idx)));
+        amd_wave_read_first_lane(compute_ptr_offset_of_groups.GetEPtrOffset(g_idx));
     const auto& ds_group_offset = compute_ptr_offset_of_groups.GetDsPtrOffset(g_idx);
 
     const long_index_t e_n_offset =
-        amd_wave_read_first_lane(int64_t(compute_ptr_offset_of_n.GetEPtrOffset(n_idx)));
+        amd_wave_read_first_lane(compute_ptr_offset_of_n.GetEPtrOffset(n_idx));
 
     __shared__ char p_shared[GridwiseGemm::GetSharedMemoryNumberOfByte()];
 
@@ -172,12 +172,12 @@ __global__ void
     else
     {
         const long_index_t a_group_offset =
-            amd_wave_read_first_lane(int64_t(compute_ptr_offset_of_groups.GetAPtrOffset(g_idx)));
+            amd_wave_read_first_lane(compute_ptr_offset_of_groups.GetAPtrOffset(g_idx));
         const long_index_t b_group_offset =
-            amd_wave_read_first_lane(int64_t(compute_ptr_offset_of_groups.GetBPtrOffset(g_idx)));
+            amd_wave_read_first_lane(compute_ptr_offset_of_groups.GetBPtrOffset(g_idx));
 
         const long_index_t a_n_offset =
-            amd_wave_read_first_lane(int64_t(compute_ptr_offset_of_n.GetAPtrOffset(n_idx)));
+            amd_wave_read_first_lane(compute_ptr_offset_of_n.GetAPtrOffset(n_idx));
 
         GridwiseGemm::template Run<HasMainKBlockLoop, InMemoryDataOperationEnum::Set>(
             p_as_grid + a_group_offset + a_n_offset,
