+++ projects/composablekernel/include/ck/tensor_operation/gpu/device/impl/device_grouped_conv_fwd_multiple_abd_xdl_cshuffle_v3.hpp
@@ -88,16 +88,16 @@ __global__ void
     const index_t n_idx = __builtin_amdgcn_readfirstlane(blockIdx.z);
 
     const long_index_t a_batch_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_groups.GetAPtrOffset(g_idx)));
+        compute_ptr_offset_of_groups.GetAPtrOffset(g_idx));
     const long_index_t b_batch_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_groups.GetBPtrOffset(g_idx)));
+        compute_ptr_offset_of_groups.GetBPtrOffset(g_idx));
     const long_index_t e_batch_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_groups.GetEPtrOffset(g_idx)));
+        compute_ptr_offset_of_groups.GetEPtrOffset(g_idx));
 
     const long_index_t a_n_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_n.GetAPtrOffset(n_idx)));
+        compute_ptr_offset_of_n.GetAPtrOffset(n_idx));
     const long_index_t e_n_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_n.GetEPtrOffset(n_idx)));
+        compute_ptr_offset_of_n.GetEPtrOffset(n_idx));
 
     __shared__ char p_shared[GridwiseGemm::GetSharedMemoryNumberOfByte()];
 
@@ -147,16 +147,16 @@ __global__ void
     const index_t n_idx = __builtin_amdgcn_readfirstlane(blockIdx.z);
 
     const long_index_t a_batch_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_groups.GetAPtrOffset(g_idx)));
+        compute_ptr_offset_of_groups.GetAPtrOffset(g_idx));
     const long_index_t b_batch_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_groups.GetBPtrOffset(g_idx)));
+        compute_ptr_offset_of_groups.GetBPtrOffset(g_idx));
     const long_index_t e_batch_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_groups.GetEPtrOffset(g_idx)));
+        compute_ptr_offset_of_groups.GetEPtrOffset(g_idx));
 
     const long_index_t a_n_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_n.GetAPtrOffset(n_idx)));
+        compute_ptr_offset_of_n.GetAPtrOffset(n_idx));
     const long_index_t e_n_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_n.GetEPtrOffset(n_idx)));
+        compute_ptr_offset_of_n.GetEPtrOffset(n_idx));
 
     // Pass two lds pointer is the key to tell compiler that ds_read/write
     // operate on different lds chunk at same time without order dependecy
