+++ projects/composablekernel/include/ck/tensor_operation/gpu/grid/gridwise_conv_fwd_multiple_d_wcnn_cshuffle.hpp
@@ -62,11 +62,11 @@ __global__ void
 
     const index_t g_idx = __builtin_amdgcn_readfirstlane(get_block_1d_id() / num_blocks_per_batch);
     const long_index_t in_batch_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_batch.GetAPtrOffset(g_idx)));
+        compute_ptr_offset_of_batch.GetAPtrOffset(g_idx));
     const long_index_t wei_batch_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_batch.GetBPtrOffset(g_idx)));
+        compute_ptr_offset_of_batch.GetBPtrOffset(g_idx));
     const long_index_t e_batch_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_batch.GetEPtrOffset(g_idx)));
+        compute_ptr_offset_of_batch.GetEPtrOffset(g_idx));
     const auto ds_batch_offset = compute_ptr_offset_of_batch.GetDsPtrOffset(g_idx);
 
     __shared__ char p_shared[GridwiseOp::BlockwiseConv::SharedMemTrait::lds_size];
@@ -364,11 +364,11 @@ __global__ void __exp_amd_wavegroup_kernel(4, 32, 256, 1, 1)
     const index_t g_idx = __builtin_amdgcn_readfirstlane(get_block_1d_id() / num_blocks_per_batch);
 
     const long_index_t in_batch_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_batch.GetAPtrOffset(g_idx)));
+        compute_ptr_offset_of_batch.GetAPtrOffset(g_idx));
     const long_index_t wei_batch_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_batch.GetBPtrOffset(g_idx)));
+        compute_ptr_offset_of_batch.GetBPtrOffset(g_idx));
     const long_index_t e_batch_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_batch.GetEPtrOffset(g_idx)));
+        compute_ptr_offset_of_batch.GetEPtrOffset(g_idx));
 
     const auto ds_batch_offset = compute_ptr_offset_of_batch.GetDsPtrOffset(g_idx);
     DsPointer p_ds_grid_grp;
@@ -458,11 +458,11 @@ __global__ void __exp_amd_wavegroup_kernel(4, 32, 512, 1, 1)
     const index_t g_idx = __builtin_amdgcn_readfirstlane(get_block_1d_id() / num_blocks_per_batch);
 
     const long_index_t in_batch_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_batch.GetAPtrOffset(g_idx)));
+        compute_ptr_offset_of_batch.GetAPtrOffset(g_idx));
     const long_index_t wei_batch_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_batch.GetBPtrOffset(g_idx)));
+        compute_ptr_offset_of_batch.GetBPtrOffset(g_idx));
     const long_index_t e_batch_offset = amd_wave_read_first_lane(
-        static_cast<int64_t>(compute_ptr_offset_of_batch.GetEPtrOffset(g_idx)));
+        compute_ptr_offset_of_batch.GetEPtrOffset(g_idx));
 
     const auto ds_batch_offset = compute_ptr_offset_of_batch.GetDsPtrOffset(g_idx);
     DsPointer p_ds_grid_grp;
